pyproject.toml
[project]
name = "weather-agent"
version = "0.1.0"
description = "Weather AI Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.28.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
